made datafilesize an unsigned long long so that max size of files to stop a run can be >4.3GB
git-svn-id: svn://orca.physics.unc.edu/Orca@8986 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Data Processing/Data Storage/ORDataFileModel.h
@@ -38,7 +38,7 @@
 {
     @private
         NSFileHandle*	filePointer;
-        unsigned long	dataFileSize;
+        unsigned long long	dataFileSize;
         NSString*		fileName;
         NSString*		statusFileName;
 
@@ -105,8 +105,8 @@
 - (NSFileHandle *)filePointer;
 - (void)setFilePointer:(NSFileHandle *)aFilePointer;
 - (void)setTitles;
-- (unsigned long)dataFileSize;
-- (void) setDataFileSize:(unsigned long)aSize;
+- (unsigned long long)dataFileSize;
+- (void) setDataFileSize:(unsigned long long)aSize;
 - (void) getDataFileSize;
 - (void) checkDiskStatus;
